+++ module/Core/src/Importer/ImportedLinksProcessor.php
@@ -11,39 +11,53 @@ use Shlinkio\Shlink\Core\ShortUrl\Helper\ShortCodeUniquenessHelperInterface;
 use Shlinkio\Shlink\Core\ShortUrl\Repository\ShortUrlRepositoryInterface;
 use Shlinkio\Shlink\Core\ShortUrl\Resolver\ShortUrlRelationResolverInterface;
 use Shlinkio\Shlink\Core\Util\DoctrineBatchHelperInterface;
+use Shlinkio\Shlink\Core\Visit\Entity\Visit;
+use Shlinkio\Shlink\Core\Visit\Repository\VisitRepositoryInterface;
 use Shlinkio\Shlink\Importer\ImportedLinksProcessorInterface;
+use Shlinkio\Shlink\Importer\Model\ImportedShlinkOrphanVisit;
 use Shlinkio\Shlink\Importer\Model\ImportedShlinkUrl;
+use Shlinkio\Shlink\Importer\Model\ImportResult;
 use Shlinkio\Shlink\Importer\Params\ImportParams;
 use Shlinkio\Shlink\Importer\Sources\ImportSource;
 use Symfony\Component\Console\Style\OutputStyle;
 use Symfony\Component\Console\Style\StyleInterface;
 use Throwable;
 
+use function Shlinkio\Shlink\Core\normalizeDate;
 use function sprintf;
 
 class ImportedLinksProcessor implements ImportedLinksProcessorInterface
 {
-    private ShortUrlRepositoryInterface $shortUrlRepo;
-
     public function __construct(
         private readonly EntityManagerInterface $em,
         private readonly ShortUrlRelationResolverInterface $relationResolver,
         private readonly ShortCodeUniquenessHelperInterface $shortCodeHelper,
         private readonly DoctrineBatchHelperInterface $batchHelper,
     ) {
-        $this->shortUrlRepo = $this->em->getRepository(ShortUrl::class);
+    }
+
+    public function process(StyleInterface $io, ImportResult $result, ImportParams $params): void
+    {
+        $io->title('Importing short URLs');
+        $this->importShortUrls($io, $result->shlinkUrls, $params);
+
+        if ($params->importOrphanVisits) {
+            $io->title('Importing orphan visits');
+            $this->importOrphanVisits($io, $result->orphanVisits);
+        }
+
+        $io->success('Data properly imported!');
     }
 
     /**
      * @param iterable<ImportedShlinkUrl> $shlinkUrls
      */
-    public function process(StyleInterface $io, iterable $shlinkUrls, ImportParams $params): void
+    private function importShortUrls(StyleInterface $io, iterable $shlinkUrls, ImportParams $params): void
     {
         $importShortCodes = $params->importShortCodes;
         $source = $params->source;
         $iterable = $this->batchHelper->wrapIterable($shlinkUrls, $source === ImportSource::SHLINK ? 10 : 100);
 
-        /** @var ImportedShlinkUrl $importedUrl */
         foreach ($iterable as $importedUrl) {
             $skipOnShortCodeConflict = static fn (): bool => $io->choice(sprintf(
                 'Failed to import URL "%s" because its short-code "%s" is already in use. Do you want to generate '
@@ -78,7 +92,9 @@ class ImportedLinksProcessor implements ImportedLinksProcessorInterface
         bool $importShortCodes,
         callable $skipOnShortCodeConflict,
     ): ShortUrlImporting {
-        $alreadyImportedShortUrl = $this->shortUrlRepo->findOneByImportedUrl($importedUrl);
+        /** @var ShortUrlRepositoryInterface $shortUrlRepo */
+        $shortUrlRepo = $this->em->getRepository(ShortUrl::class);
+        $alreadyImportedShortUrl = $shortUrlRepo->findOneByImportedUrl($importedUrl);
         if ($alreadyImportedShortUrl !== null) {
             return ShortUrlImporting::fromExistingShortUrl($alreadyImportedShortUrl);
         }
@@ -106,5 +122,30 @@ class ImportedLinksProcessor implements ImportedLinksProcessorInterface
         }
 
         return $this->shortCodeHelper->ensureShortCodeUniqueness($shortUrl, false);
+    }
+
+    /**
+     * @param iterable<ImportedShlinkOrphanVisit> $orphanVisits
+     */
+    private function importOrphanVisits(StyleInterface $io, iterable $orphanVisits): void
+    {
+        $iterable = $this->batchHelper->wrapIterable($orphanVisits, 100);
+
+        /** @var VisitRepositoryInterface $visitRepo */
+        $visitRepo = $this->em->getRepository(Visit::class);
+        $mostRecentOrphanVisit = $visitRepo->findMostRecentOrphanVisit();
+
+        $importedVisits = 0;
+        foreach ($iterable as $importedOrphanVisit) {
+            // Skip visits which are older than the most recent already imported visit's date
+            if ($mostRecentOrphanVisit?->getDate()->gte(normalizeDate($importedOrphanVisit->date))) {
+                continue;
+            }
+
+            $this->em->persist(Visit::fromOrphanImport($importedOrphanVisit));
+            $importedVisits++;
+        }
+
+        $io->text(sprintf('<info>Imported %s</info> orphan visits.', $importedVisits));
     }
 }