+++ module/Core/src/Util/DoctrineBatchHelperInterface.php
@@ -6,5 +6,10 @@ namespace Shlinkio\Shlink\Core\Util;
 
 interface DoctrineBatchHelperInterface
 {
+    /**
+     * @template T
+     * @param iterable<T> $resultSet
+     * @return iterable<T>
+     */
     public function wrapIterable(iterable $resultSet, int $batchSize): iterable;
 }