+++ module/Core/test/Importer/ImportedLinksProcessorTest.php
@@ -17,8 +17,12 @@ use Shlinkio\Shlink\Core\ShortUrl\Repository\ShortUrlRepositoryInterface;
 use Shlinkio\Shlink\Core\ShortUrl\Resolver\SimpleShortUrlRelationResolver;
 use Shlinkio\Shlink\Core\Util\DoctrineBatchHelperInterface;
 use Shlinkio\Shlink\Core\Visit\Entity\Visit;
+use Shlinkio\Shlink\Core\Visit\Model\Visitor;
+use Shlinkio\Shlink\Core\Visit\Repository\VisitRepositoryInterface;
+use Shlinkio\Shlink\Importer\Model\ImportedShlinkOrphanVisit;
 use Shlinkio\Shlink\Importer\Model\ImportedShlinkUrl;
 use Shlinkio\Shlink\Importer\Model\ImportedShlinkVisit;
+use Shlinkio\Shlink\Importer\Model\ImportResult;
 use Shlinkio\Shlink\Importer\Params\ImportParams;
 use Shlinkio\Shlink\Importer\Sources\ImportSource;
 use stdClass;
@@ -27,6 +31,7 @@ use Symfony\Component\Console\Style\StyleInterface;
 use function count;
 use function Functional\contains;
 use function Functional\some;
+use function sprintf;
 use function str_contains;
 
 class ImportedLinksProcessorTest extends TestCase
@@ -41,7 +46,6 @@ class ImportedLinksProcessorTest extends TestCase
     {
         $this->em = $this->createMock(EntityManagerInterface::class);
         $this->repo = $this->createMock(ShortUrlRepositoryInterface::class);
-        $this->em->method('getRepository')->with(ShortUrl::class)->willReturn($this->repo);
 
         $this->shortCodeHelper = $this->createMock(ShortCodeUniquenessHelperInterface::class);
         $batchHelper = $this->createMock(DoctrineBatchHelperInterface::class);
@@ -67,6 +71,7 @@ class ImportedLinksProcessorTest extends TestCase
         ];
         $expectedCalls = count($urls);
 
+        $this->em->method('getRepository')->with(ShortUrl::class)->willReturn($this->repo);
         $this->repo->expects($this->exactly($expectedCalls))->method('findOneByImportedUrl')->willReturn(null);
         $this->shortCodeHelper->expects($this->exactly($expectedCalls))
                               ->method('ensureShortCodeUniqueness')
@@ -76,7 +81,7 @@ class ImportedLinksProcessorTest extends TestCase
         );
         $this->io->expects($this->exactly($expectedCalls))->method('text')->with($this->isType('string'));
 
-        $this->processor->process($this->io, $urls, $this->buildParams());
+        $this->processor->process($this->io, ImportResult::withShortUrls($urls), $this->buildParams());
     }
 
     /** @test */
@@ -88,6 +93,7 @@ class ImportedLinksProcessorTest extends TestCase
             new ImportedShlinkUrl(ImportSource::BITLY, 'baz', [], Chronos::now(), null, 'baz', null),
         ];
 
+        $this->em->method('getRepository')->with(ShortUrl::class)->willReturn($this->repo);
         $this->repo->expects($this->exactly(3))->method('findOneByImportedUrl')->willReturn(null);
         $this->shortCodeHelper->expects($this->exactly(3))->method('ensureShortCodeUniqueness')->willReturn(true);
         $this->em->expects($this->exactly(3))->method('persist')->with(
@@ -99,7 +105,7 @@ class ImportedLinksProcessorTest extends TestCase
         });
         $textCalls = $this->setUpIoText('<comment>Skipped</comment>. Reason: Whatever error', '<info>Imported</info>');
 
-        $this->processor->process($this->io, $urls, $this->buildParams());
+        $this->processor->process($this->io, ImportResult::withShortUrls($urls), $this->buildParams());
 
         self::assertEquals(2, $textCalls->importedCount);
         self::assertEquals(1, $textCalls->skippedCount);
@@ -116,6 +122,7 @@ class ImportedLinksProcessorTest extends TestCase
             new ImportedShlinkUrl(ImportSource::BITLY, 'baz3', [], Chronos::now(), null, 'baz3', null),
         ];
 
+        $this->em->method('getRepository')->with(ShortUrl::class)->willReturn($this->repo);
         $this->repo->expects($this->exactly(count($urls)))->method('findOneByImportedUrl')->willReturnCallback(
             fn (ImportedShlinkUrl $url): ?ShortUrl
                 => contains(['foo', 'baz2', 'baz3'], $url->longUrl) ? ShortUrl::fromImport($url, true) : null,
@@ -124,7 +131,7 @@ class ImportedLinksProcessorTest extends TestCase
         $this->em->expects($this->exactly(2))->method('persist')->with($this->isInstanceOf(ShortUrl::class));
         $textCalls = $this->setUpIoText();
 
-        $this->processor->process($this->io, $urls, $this->buildParams());
+        $this->processor->process($this->io, ImportResult::withShortUrls($urls), $this->buildParams());
 
         self::assertEquals(2, $textCalls->importedCount);
         self::assertEquals(3, $textCalls->skippedCount);
@@ -141,6 +148,7 @@ class ImportedLinksProcessorTest extends TestCase
             new ImportedShlinkUrl(ImportSource::BITLY, 'baz3', [], Chronos::now(), null, 'baz3', 'bar'),
         ];
 
+        $this->em->method('getRepository')->with(ShortUrl::class)->willReturn($this->repo);
         $this->repo->expects($this->exactly(count($urls)))->method('findOneByImportedUrl')->willReturn(null);
         $this->shortCodeHelper->expects($this->exactly(7))->method('ensureShortCodeUniqueness')->willReturnCallback(
             fn ($_, bool $hasCustomSlug) => ! $hasCustomSlug,
@@ -151,7 +159,7 @@ class ImportedLinksProcessorTest extends TestCase
         });
         $textCalls = $this->setUpIoText('Error');
 
-        $this->processor->process($this->io, $urls, $this->buildParams());
+        $this->processor->process($this->io, ImportResult::withShortUrls($urls), $this->buildParams());
 
         self::assertEquals(2, $textCalls->importedCount);
         self::assertEquals(3, $textCalls->skippedCount);
@@ -167,6 +175,7 @@ class ImportedLinksProcessorTest extends TestCase
         int $amountOfPersistedVisits,
         ?ShortUrl $foundShortUrl,
     ): void {
+        $this->em->method('getRepository')->with(ShortUrl::class)->willReturn($this->repo);
         $this->repo->expects($this->once())->method('findOneByImportedUrl')->willReturn($foundShortUrl);
         $this->shortCodeHelper->expects($this->exactly($foundShortUrl === null ? 1 : 0))
                               ->method('ensureShortCodeUniqueness')
@@ -176,7 +185,7 @@ class ImportedLinksProcessorTest extends TestCase
         )->with($this->callback(fn (object $arg) => $arg instanceof ShortUrl || $arg instanceof Visit));
         $this->io->expects($this->once())->method('text')->with($this->stringContains($expectedOutput));
 
-        $this->processor->process($this->io, [$importedUrl], $this->buildParams());
+        $this->processor->process($this->io, ImportResult::withShortUrls([$importedUrl]), $this->buildParams());
     }
 
     public function provideUrlsWithVisits(): iterable
@@ -219,9 +228,69 @@ class ImportedLinksProcessorTest extends TestCase
         ];
     }
 
-    private function buildParams(): ImportParams
+    /**
+     * @param iterable<ImportedShlinkOrphanVisit> $visits
+     * @test
+     * @dataProvider provideOrphanVisits
+     */
+    public function properAmountOfOrphanVisitsIsImported(
+        bool $importOrphanVisits,
+        iterable $visits,
+        ?Visit $lastOrphanVisit,
+        int $expectedImportedVisits,
+    ): void {
+        $this->io->expects($this->exactly($importOrphanVisits ? 2 : 1))->method('title');
+        $this->io->expects($importOrphanVisits ? $this->once() : $this->never())->method('text')->with(
+            sprintf('<info>Imported %s</info> orphan visits.', $expectedImportedVisits),
+        );
+
+        $visitRepo = $this->createMock(VisitRepositoryInterface::class);
+        $visitRepo->expects($importOrphanVisits ? $this->once() : $this->never())->method(
+            'findMostRecentOrphanVisit',
+        )->willReturn($lastOrphanVisit);
+        $this->em->expects($importOrphanVisits ? $this->once() : $this->never())->method('getRepository')->with(
+            Visit::class,
+        )->willReturn($visitRepo);
+        $this->em->expects($importOrphanVisits ? $this->exactly($expectedImportedVisits) : $this->never())->method(
+            'persist',
+        )->with($this->isInstanceOf(Visit::class));
+
+        $this->processor->process(
+            $this->io,
+            ImportResult::withShortUrlsAndOrphanVisits([], $visits),
+            $this->buildParams($importOrphanVisits),
+        );
+    }
+
+    public function provideOrphanVisits(): iterable
     {
-        return ImportSource::BITLY->toParamsWithCallableMap(['import_short_codes' => static fn () => true]);
+        yield 'import orphan disable without visits' => [false, [], null, 0];
+        yield 'import orphan enabled without visits' => [true, [], null, 0];
+        yield 'import orphan disabled with visits' => [false, [
+            new ImportedShlinkOrphanVisit('', '', Chronos::now(), '', '', null),
+        ], null, 0];
+        yield 'import orphan enabled with visits' => [true, [
+            new ImportedShlinkOrphanVisit('', '', Chronos::now(), '', '', null),
+            new ImportedShlinkOrphanVisit('', '', Chronos::now(), '', '', null),
+            new ImportedShlinkOrphanVisit('', '', Chronos::now(), '', '', null),
+            new ImportedShlinkOrphanVisit('', '', Chronos::now(), '', '', null),
+            new ImportedShlinkOrphanVisit('', '', Chronos::now(), '', '', null),
+        ], null, 5];
+        yield 'existing orphan visit' => [true, [
+            new ImportedShlinkOrphanVisit('', '', Chronos::now()->subDays(3), '', '', null),
+            new ImportedShlinkOrphanVisit('', '', Chronos::now()->subDays(2), '', '', null),
+            new ImportedShlinkOrphanVisit('', '', Chronos::now()->addDay(), '', '', null),
+            new ImportedShlinkOrphanVisit('', '', Chronos::now()->addDay(), '', '', null),
+            new ImportedShlinkOrphanVisit('', '', Chronos::now()->addDay(), '', '', null),
+        ], Visit::forBasePath(Visitor::botInstance()), 3];
+    }
+
+    private function buildParams(bool $importOrphanVisits = false): ImportParams
+    {
+        return ImportSource::BITLY->toParamsWithCallableMap([
+            ImportParams::IMPORT_SHORT_CODES_PARAM => static fn () => true,
+            ImportParams::IMPORT_ORPHAN_VISITS_PARAM => static fn () => $importOrphanVisits,
+        ]);
     }
 
     public function setUpIoText(string $skippedText = 'Skipped', string $importedText = 'Imported'): stdClass